refactor: 공통 API 호출부 다른 파일로 분리
